requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
httpx>=0.25.0
python-multipart>=0.0.6

